Refactored update handling
