+++ holoviews/plotting/bokeh/plot.py
@@ -79,48 +79,8 @@ class BokehPlot(DimensionedPlot):
     backend = 'bokeh'
 
     @property
-    def document(self):
-        return self._document
-
-    @property
     def id(self):
         return self.root.ref['id'] if self.root else None
-
-    @property
-    def root(self):
-        if self._root:
-            return self._root
-        elif 'plot' in self.handles and self.top_level:
-            return self.state
-        else:
-            return None
-
-    @document.setter
-    def document(self, doc):
-        if (doc and hasattr(doc, 'on_session_destroyed') and
-            self.root is self.handles.get('plot') and not isinstance(self, AdjointLayoutPlot)):
-            doc.on_session_destroyed(self._session_destroy)
-            if self._document:
-                if isinstance(self._document._session_destroyed_callbacks, set):
-                    self._document._session_destroyed_callbacks.discard(self._session_destroy)
-                else:
-                    self._document._session_destroyed_callbacks.pop(self._session_destroy, None)
-
-        self._document = doc
-        if self.subplots:
-            for plot in self.subplots.values():
-                if plot is not None:
-                    plot.document = doc
-
-
-    def _session_destroy(self, session_context):
-        self.cleanup()
-
-    def __init__(self, *args, **params):
-        root = params.pop('root', None)
-        super(BokehPlot, self).__init__(*args, **params)
-        self._document = None
-        self._root = root
 
 
     def get_data(self, element, ranges, style):
@@ -175,36 +135,6 @@ class BokehPlot(DimensionedPlot):
             cb_streams = [s for _, s in group]
             cbs.append(cb(self, cb_streams, source))
         return cbs
-
-    def refresh(self, **kwargs):
-        if self.renderer.mode == 'server' and curdoc() is not self.document:
-            # If we do not have the Document lock, schedule refresh as callback
-            self.document.add_next_tick_callback(self.refresh)
-        else:
-            super(BokehPlot, self).refresh(**kwargs)
-
-    def push(self):
-        """
-        Pushes updated plot data via the Comm.
-        """
-        if self.renderer.mode == 'server':
-            return
-        if self.comm is None:
-            raise Exception('Renderer does not have a comm.')
-
-        if self._root and 'embedded' in self._root.tags:
-            # Allows external libraries to prevent comm updates
-            return
-
-        msg = self.renderer.diff(self, binary=True)
-        if msg is None:
-            return
-        self.comm.send(msg.header_json)
-        self.comm.send(msg.metadata_json)
-        self.comm.send(msg.content_json)
-        for header, payload in msg.buffers:
-            self.comm.send(json.dumps(header))
-            self.comm.send(buffers=[payload])
 
 
     def set_root(self, root):
